# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PocketSync API
  description: |
    PocketSync REST API.

    Production: `https://pocketsync.onrender.com`

    Auth uses HttpOnly cookies (`accessToken`, `refreshToken`). Send cookies on protected
    routes (`credentials: 'include'` / `withCredentials: true`).

    Sign up: register → verify-otp → login.
    Password reset: forgot-password → reset-password.
    BVN onboarding: status → bvn/submit → bvn/verify-otp → bvn/accounts → bvn/connect.

    BVN, SMS, external transfers, and bill pay are simulated in this build.
    OTPs log to the server console; `devOtp` may appear in dev responses.

    Amounts are in NGN. Credits are positive; debits are negative in transaction objects.
  version: 1.0.0
  contact:
    name: PocketSync Backend Team
servers:
  - url: https://pocketsync.onrender.com
    description: Production (Render)
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Uses `refreshToken` cookie (path-scoped to this endpoint).
        Rotates both tokens on success.
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token refreshed
        '401':
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/accounts/link:
    post:
      tags: [Accounts]
      summary: Link a bank account
      operationId: linkAccount
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkAccountRequest'
            example:
              institution: GTBank
              mockAccountRef: "1234567890"
      responses:
        '201':
          description: Account linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Institution already linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/accounts:
    get:
      tags: [Accounts]
      summary: List linked accounts
      operationId: getAccounts
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Active linked accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/accounts/{accountId}:
    delete:
      tags: [Accounts]
      summary: Disconnect (soft-delete) a linked account
      operationId: disconnectAccount
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/accountId'
      responses:
        '200':
          description: Account disconnected
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: GTBank account disconnected successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly cookie set by POST /api/v1/auth/login
  parameters:
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId of linked account
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorised — please log in
    Forbidden:
      description: Resource exists but not owned by user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          description: Present in development only
...
    LinkableInstitution:
      type: string
      enum: [GTBank, Access Bank, Kuda, Opay, Moniepoint]
    AccountType:
      type: string
      enum: [current, savings, wallet, business]
...
    LinkedAccount:
      type: object
      properties:
        id:
          type: string
        institution:
          $ref: '#/components/schemas/LinkableInstitution'
        maskedAccountNumber:
          type: string
          example: "****4471"
        balance:
          type: number
          description: Balance in NGN
        currency:
          type: string
          example: NGN
        accountType:
          $ref: '#/components/schemas/AccountType'
        linkedAt:
          type: string
          format: date-time
...
    LinkAccountRequest:
      type: object
      required: [institution]
      properties:
        institution:
          $ref: '#/components/schemas/LinkableInstitution'
        mockAccountRef:
          type: string
          description: Optional. Last 4 digits used for masked number.
    LinkAccountResponse:
      type: object
      properties:
        message:
          type: string
        account:
          $ref: '#/components/schemas/LinkedAccount'
    AccountsListResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/LinkedAccount'
...